feat: metadata Optimization
