+++ src/app/layout.js
@@ -14,8 +14,23 @@ const geistMono = Geist_Mono({
 });
 
 export const metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: {
+    template: "%s - Think Twice",
+    default: "Think Twice",
+  },
+  description:
+    "a social media application designed to foster mindful consumption and empower users to make smarter purchasing decisions. Think Twice aims to reduce impulsive buying by providing a platform for users to share potential purchases, receive feedback from a supportive community, and collectively evaluate spending choices.",
+  authors: [{ name: "Devscale", url: "https://devscale.id/" }],
+  creator: "Lumiva Team",
+  keywords: [
+    "think twice",
+    "social media",
+    "mindful consumption",
+    "smart purchasing",
+    "community feedback",
+    "impulsive buying",
+    "spending choices",
+  ],
 };
 
 export default function RootLayout({ children }) {
